src/requirements.txt
# Core dependencies
networkx>=2.6.0
pydantic>=1.10.0
toml>=0.10.2
pyyaml>=6.0

# Optional dependencies for enhanced features
matplotlib>=3.5.0  # For visualization
plotly>=5.0.0      # For interactive charts
pandas>=1.3.0      # For data analysis
numpy>=1.20.0      # For numerical operations

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
